add custom records

+++ decoder.go
@@ -103,7 +103,7 @@ func (d *decoder) decodeRecord(rec *Record) error {
 	}
 
 	// skip crc checking if the record type is crcType
-	if rec.Type != crcType {
+	if rec.Type != int64(crcType) {
 		d.crc.Write(rec.Data)
 		if err := rec.Validate(d.crc.Sum32()); err != nil {
 			if d.isTornEntry(data) {
